% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooi_check_availability.R
\name{ooi_check_availability}
\alias{ooi_check_availability}
\title{Get deployment and annotation data associated with an OOI site, node, and instrument.}
\usage{
ooi_check_availability(site, node, instrument, user, token)
}
\arguments{
\item{site}{An eight (8) character designator for OOI sites.}

\item{node}{A five (5) character designator for OOI nodes. Alternatively, a simplified string for OOI sites that is greater than 5 characters. grep allows for use of substrings.}

\item{instrument}{A twelve (12) character designator for the OOI instrument on the associated site and node. grep allows for use of substrings.}

\item{user}{Your OOI API username.}

\item{token}{Your OOI API token.}
}
\value{
A list of sublists. The deployments sublist offers start and stop deployments times for the given asset. The annotations sublists offers annotations for the given site,node, and instrument. Used together, a user can determine if data exists for the time period they are interested in.
}
\description{
Get deployment and annotation data associated with an OOI site, node, and instrument.
}
\examples{
availability = ooi_check_availability('CE01ISSM','SEAFLOOR','CTD','OOI-API-USERNAME-HERE','OOI-API-TOKEN-HERE')
}
