% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooi_download_data.R
\name{ooi_download_data}
\alias{ooi_download_data}
\title{Take dodsC OpenDAP URLs, converts to fileServer URLs, and downloads data to a given directory as a NetCDF.}
\usage{
ooi_download_data(remote, directory = getwd())
}
\arguments{
\item{remote}{A list of OpenDAP urls where OOI data is located. Must all be from the same instrument type.}

\item{directory}{Default is to use the current working directory. If a directory is set in an R readable fashion, data will be downloaded to that directory.}
}
\value{
A list of full filepath filenames where the data was downloaded. Further utilized by the ooi_get_data() function.
}
\description{
Take dodsC OpenDAP URLs, converts to fileServer URLs, and downloads data to a given directory as a NetCDF.
}
\examples{
local = ooi_download_data(remote,'C:/Users/Ian/Desktop') #Downloads data to my desktop.
}
