% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooi_get_data.R
\name{ooi_get_data}
\alias{ooi_get_data}
\title{Pull OOI data from local files (or OpenDAP urls) into the workspace.}
\usage{
ooi_get_data(local, simplify_data = TRUE)
}
\arguments{
\item{local}{A list of local fullpath filenames to pull data from. Can be created from ooi_download_data() or manually input.}

\item{simplify_data}{A flag that will drop annoying variables from ingested data if set to TRUE. If set to FALSE, all variables are brought into the workspace. If set to TRUE, it will pull from the curated list of variables.}
}
\value{
A list of two sublists. Sublist 1 is a dataframe of data if it is 1D and can be combined or a sublist of more sublists if only one dataset if imported, or if data is 2D.
Further sublists are named by the filename, but can also be called by position. Within the filename sublists exist data as lists. Did that make sense?
Sublist 2 is a dataframe of imported variables, associated units, and variable descriptions.
}
\description{
NOTE: The ncdf4 has an issue with accessing OpenDAP urls on Windows machines, thus data must be downloaded before reading it in. If using a MacOS or Linux machine, local parameter can be a list of OpenDAP urls.
}
\examples{
lol = ooi_get_data(local,simplify_data = TRUE)
}
