% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooi_get_location.R
\name{ooi_get_location}
\alias{ooi_get_location}
\title{Check data request made to OOINet and return OpenDAP urls when complete.}
\usage{
ooi_get_location(response, drop_paired = TRUE)
}
\arguments{
\item{response}{The response object made through the ooi_submit_request function.}

\item{drop_paired}{A TRUE/FALSE flag that removes paired datasets from the return. Setting this to TRUE makes it easier to concatenate multiple files.}
}
\value{
A list of OpenDAP urls where the data is located on the OOI Thredds server as dodsC NetCDFs.
}
\description{
Check data request made to OOINet and return OpenDAP urls when complete.
}
\examples{
remote = ooi_get_location(response,drop_paired = TRUE)
}
