% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooi_site_info.R
\name{ooi_site_info}
\alias{ooi_site_info}
\title{Get a lists of lists of site information.}
\usage{
ooi_site_info(site, user, token)
}
\arguments{
\item{site}{An eight (8) character designator for an OOI site.}

\item{user}{Your OOI API username.}

\item{token}{Your OOI API token.}
}
\value{
A cascading list of lists. The first sublist layer consists of nodes. The layer in each node consists of the instruments connected to that node. Each layer under each instrument contains instrument metadata.
}
\description{
Get a lists of lists of site information.
}
\examples{
info = ooi_site_info(site = 'CE01ISSP', user = 'OOI-API-USERNAME-HERE', token = 'OOI-API-TOKEN-HERE')
info = ooi_site_info(site = 'CE05MOAS','OOI-API-USERNAME-HERE','OOI-API-TOKEN-HERE')
}
